pyproject.toml
[tool.poetry]
name = "inoa"
version = "0.1.0"
description = ""
authors = ["Vitor Toledo <vitor.toledo@grupomateus.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
celery = "^5.4.0"
django-celery-beat = "^2.6.0"
gevent = "^24.2.1"
djangorestframework = "^3.15.2"
markdown = "^3.6"
django-filter = "^24.2"
sqlalchemy = "^2.0.31"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
